// src/station_information.d.ts
// Generated by https://quicktype.io

export interface StationInformation {
  last_updated: number;
  ttl: number;
  data: Data;
}

export interface Data {
  stations: Station[];
}

export interface Station {
  station_id: string;
  external_id: string;
  name: string;
  short_name: string;
  lat: number;
  lon: number;
  region_id: number;
  rental_methods: RentalMethod[];
  capacity: number;
  rental_url: string;
  electric_bike_surcharge_waiver: boolean;
  eightd_has_key_dispenser: boolean;
  eightd_station_services?: EightdStationService[];
  has_kiosk: boolean;
}

export interface EightdStationService {
  id: string;
  service_type: ServiceType;
  bikes_availability: BikesAvailability;
  docks_availability: DocksAvailability;
  name: Name;
  description: string;
  schedule_description: string;
  link_for_more_info: string;
}

export enum BikesAvailability {
  Unlimited = 'UNLIMITED',
}

export enum DocksAvailability {
  None = 'NONE',
}

export enum Name {
  ValetService = 'Valet Service',
}

export enum ServiceType {
  AttendedService = 'ATTENDED_SERVICE',
}

export enum RentalMethod {
  Creditcard = 'CREDITCARD',
  Key = 'KEY',
}